+++ src/resources/testng.xml
@@ -6,7 +6,7 @@
         <classes>
             <class name="com.automationpractice.tests.CreateAccountPageTest" />
             <class name="com.automationpractice.tests.HomePageTest" />
-            <class name="com.automationpractice.tests.LoginPageTest" />
+            <class name="com.automationpractice.tests.login.PositiveLoginTest" />
         </classes>
     </test>
 </suite>
